Mark TestEvaluatorLifecycle as integration test
